CRUD Operation

+++ 03-practicing/includes/helpers/storage.php
@@ -20,12 +20,23 @@ if (!function_exists('storage')) {
     }
 }
 
-if (!function_exists('delete_file')) {
-    function delete_file($path)
+
+if (!function_exists('storage_url')) {
+    function storage_url(string $path=null):string
     {
-        if (file_exists($path)) {
+        return !empty($path) ? url('storage/'.$path) : '';
+    }
+}
+
+
+if (!function_exists('delete_file')) {
+    function delete_file($to_path)
+    {
+        $path =rtrim(config('files.storage_files_path'),'/').'/'.ltrim($to_path ,'/');
+        if (file_exists($path) && is_file($path)) {
             return unlink($path);
         }
+        return false;
     }
 }
 
@@ -39,27 +50,25 @@ if (!function_exists('remove_folder')) {
     }
 }
 
-
 if (!function_exists('store_file')) {
     function store_file(array $from, string $to): bool|string
     {
         // Check if the temporary file name is set
         if (isset($from['tmp_name']) && !empty($from['tmp_name'])) {
-            $to_path = '/' . ltrim($to, '/');
-            $path = config('files.storage_files_path') . $to_path;
-
-            $ex_path = explode('/', $path);
-            $end_file = end($ex_path);
-            $check_path = str_replace($end_file, '', $path);
+            $file_info = file_ext($from);  // Get the file extension and hash name
+            $to_path = rtrim($to, '/') . '/' . $file_info['hash_name'];  // Append hash name to the path
+            $path = config('files.storage_files_path') . '/' . ltrim($to_path, '/');
+            
+            $directory_path = dirname($path);
 
             // Check if the directory needs to be created
-            if (!empty($check_path) && !is_dir($check_path)) {
-                mkdir($check_path, 0777, true);
+            if (!is_dir($directory_path)) {
+                mkdir($directory_path, 0777, true);
             }
 
             // Move the uploaded file to the specified path
             if (move_uploaded_file($from['tmp_name'], $path)) {
-                return $to;
+                return $to_path;
             }
         }
         return false;
@@ -69,15 +78,16 @@ if (!function_exists('store_file')) {
 
 
 
-if (!function_exists('file-ext')) {
+
+if (!function_exists('file_ext')) {
     function file_ext(array $file_name): array
     {
         if (!empty($file_name['name'])) {
             $fext = explode('.', $file_name['name']);
             $file_ext = end($fext);
-            $hash_name = md5(10) .rand(000,999). '.' . $file_ext;
+            $hash_name = md5(uniqid()) . rand(000, 999) . '.' . $file_ext;
             return [
-                'name' => $file_name,
+                'name' => $file_name['name'],
                 'hash_name' => $hash_name,
                 'ext' => $file_ext,
             ];
@@ -90,3 +100,4 @@ if (!function_exists('file-ext')) {
         }
     }
 }
+
